Remove TTD measurement for delayed stanzas
This prevents us from measuring TTD for stanzas that come from mod_offline.
This would not be real TTD. However, if users don't check whether the receiver
for a message is online, some messages may be sent in vain, and not be measured.
For now, there is no mechanism in place to check that.