Move outdated scenarios to a separate directory
The main goal is to have an immediate distinciton between the scenarios
confirmed to work with the latest MongooseIM and the rest.

If an old scenario is updated to work with the latest MIM again,
it can be moved back.

For the non-outdated scenarios, we will try to keep them updated
by keeping using them in our regression load tests.